requirements.txt
google-generativeai==0.5.3
python-dotenv==1.0.1
PyMuPDF==1.24.4
streamlit==1.34.0